feat: love toggle on search results
